+++ src/ts/PlotBoilerplate.ts
@@ -82,7 +82,8 @@
  * @modified 2023-02-10 Fixing an issue of the `style.position` setting when `fitToParent=true` from `absolute` to `static` (default).
  * @modified 2023-02-10 Cleaning up most type errors in the main class (mostly null checks).
  * @modified 2023-02-10 Adding `enableZoom` and `enablePan` (both default true) to have the option to disable these functions.
- * @version  1.17.2
+ * @modified 2023-09-29 Adding proper dicionary key and value types to the params of `PlotBoilerplate.utils.safeMergeByKeys` (was `object` before).
+ * @version  1.17.3
  *
  * @file PlotBoilerplate
  * @fileoverview The main class.
@@ -399,7 +400,7 @@ export class PlotBoilerplate {
       visible: true
     };
 
-    if (typeof config.canvas == "undefined") {
+    if (typeof config.canvas === "undefined") {
       throw "No canvas specified.";
     }
 
@@ -557,7 +558,10 @@ export class PlotBoilerplate {
     this.grid = new Grid(new Vertex(0, 0), new Vertex(50, 50));
     this.canvasSize = { width: PlotBoilerplate.DEFAULT_CANVAS_WIDTH, height: PlotBoilerplate.DEFAULT_CANVAS_HEIGHT };
     const canvasElement: Element =
-      typeof config.canvas == "string" ? (document.querySelector(config.canvas) as Element) : config.canvas;
+      typeof config.canvas === "string" ? (document.querySelector(config.canvas) as Element) : config.canvas;
+    if (typeof canvasElement === "undefined") {
+      throw `Cannot initialize PlotBoilerplate with a null canvas (element "${config.canvas} not found).`;
+    }
     // Which renderer to use: Canvas2D, WebGL (experimental) or SVG?
     if (canvasElement.tagName.toLowerCase() === "canvas") {
       this.canvas = canvasElement as HTMLCanvasElement;
@@ -2233,7 +2237,45 @@ export class PlotBoilerplate {
      * @param {Object} extension
      * @return {Object} base extended by the new attributes.
      **/
-    safeMergeByKeys: (base: Object, extension: Object): Object => {
+    safeMergeByKeys: <KeyType extends string | number | symbol, ValueType>(
+      base: Record<KeyType, ValueType>,
+      extension: Record<KeyType, ValueType>
+    ): Record<KeyType, ValueType> => {
+      for (var k in extension) {
+        if (!extension.hasOwnProperty(k)) {
+          continue;
+        }
+        if (base.hasOwnProperty(k)) {
+          const typ = typeof base[k];
+          const extVal = extension[k];
+          try {
+            if (typ == "boolean") {
+              if (typeof extVal === "string") base[k] = Boolean(!!JSON.parse(extVal)) as unknown as ValueType;
+              else base[k] = extVal as ValueType;
+            } else if (typ == "number") {
+              if (typeof extVal === "string") base[k] = Number(JSON.parse(extVal) * 1) as unknown as ValueType;
+              else base[k] = extension[k] as ValueType;
+            } else if (typ == "function" && typeof extVal == "function") {
+              base[k] = extension[k];
+            } else {
+              // Probably a sting
+              base[k] = extension[k];
+            }
+          } catch (e) {
+            console.error("error in key ", k, extVal, e);
+          }
+        } else {
+          base[k] = extension[k];
+        }
+      }
+      return base;
+    },
+
+    /*
+    __safeMergeByKeys: <KeyType extends string | number | symbol, ValueType extends boolean | number | string | Function>(
+      base: Record<KeyType, ValueType>,
+      extension: Record<KeyType, string>
+    ): Record<KeyType, ValueType> => {
       for (var k in extension) {
         if (!extension.hasOwnProperty(k)) continue;
         if (base.hasOwnProperty(k)) {
@@ -2252,6 +2294,7 @@ export class PlotBoilerplate {
       }
       return base;
     },
+    *()
 
     /**
      * A helper function to scale elements (usually the canvas) using CSS.