+++ src/ts/interfaces/core.ts
@@ -18,6 +18,7 @@
  * @modified 2022-10-25 Added `origin` param to the `DrawConfig` interface.
  * @modified 2022-11-23 Added `drawRaster` to the `Config` interface.
  * @modified 2023-02-10 All non-function attributes of the `Config` interface are now mandatory.
+ * @modified 2023-09-29 Added the `randomPoint(...)` function declaration to the IBounds interface.
  **/
 
 import { Vertex } from "../Vertex";
@@ -67,6 +68,16 @@ export interface IBounds {
   min: XYCoords;
   max: XYCoords;
   getCenter(): Vertex;
+  /**
+   * Generate a random point inside this bounds object. Safe areas at the border to avoid
+   * included.
+   *
+   * @method randomPoint
+   * @param {horizontalSafeArea} - (optional) The horizonal (left and right) safe area. No vertex will be created here. Can be used as percent in (0.0 ... 0.1) interval.
+   * @param {verticalSafeArea} - (optional) The vertical (top and bottom) safe area. No vertex will be created here. Can be used as percent in (0.0 ... 0.1) interval
+   * @returns {Vertex} A pseudo random point inside these bounds.
+   */
+  randomPoint: (horizontalSafeArea?: number, verticalSafeArea?: number) => Vertex;
 }
 
 /**
@@ -101,8 +112,9 @@ export interface CanvasWrapper {
 /**
  * The config that's used by PB.
  */
-export interface Config {
-  canvas: HTMLCanvasElement | string; //  Your canvas element in the DOM (required).
+export interface Config
+  extends Record<string, boolean | number | string | Function | HTMLCanvasElement | SVGElement | undefined> {
+  canvas: HTMLCanvasElement | SVGElement | string; //  Your canvas element in the DOM (required).
   fullSize: boolean; // If set to true the canvas will gain full window size.
   fitToParent: boolean; // If set to true the canvas will gain the size of its parent container (overrides fullSize).
   scaleX: number; // The initial x-zoom. Default is 1.0.